extracted some in the helper thingy
