+++ fetchers/agora.rb
@@ -2,43 +2,15 @@ require_relative "../main"
 
 module Fetchers
   class Agora
+    include Fetchers::PagesWithIntegerIdsInUrl
+
     FEED_URL  = "http://agora.md/rss/news"
 
     def initialize(storage=LocalStorageFactory.agora)
       @storage = storage
     end
 
-    def run
-      puts "Fetching Timpul. Most recent: #{most_recent_id}. Last fetched: #{latest_stored_id}."
-
-      if all_pages_are_fetched?
-        puts "Nothing to fetch for Timpul"
-        return
-      end
-
-      latest_stored_id.upto(most_recent_id) do |id|
-        fetch_single(id)
-        progressbar.increment!
-      end
-    end
-
   private
-
-    def all_pages_are_fetched?
-      latest_stored_id == most_recent_id
-    end
-
-    def latest_stored_id
-      @storage.latest_page_id
-    end
-
-    def save(page, id)
-      @storage.save(page, id)
-    end
-
-    def most_recent_id
-      @most_recent_id ||= fetch_most_recent_id
-    end
 
     def fetch_most_recent_id
       RestClient.get(FEED_URL)
@@ -58,12 +30,7 @@ module Fetchers
     end
 
     def fetch_single(id)
-      page = SmartFetcher.fetch(link(id))
-      save(page, id) if valid?(page)
-    end
-
-    def progressbar
-      @progressbar ||= ProgressBar.new(most_recent_id - latest_stored_id, :bar, :counter, :rate, :eta)
+      SmartFetcher.fetch(link(id))
     end
   end
 end
