+++ fetchers/pro_tv.rb
@@ -2,6 +2,8 @@ require_relative "../main"
 
 module Fetchers
   class ProTv
+    include Fetchers::PagesWithIntegerIdsInUrl
+
     MAIN_PAGE = "http://protv.md"
     LATEST_NEWS = "http://protv.md/export/featured/json"
 
@@ -9,37 +11,7 @@ module Fetchers
       @storage = storage
     end
 
-    def run
-      puts "Fetching Protv. Most recent: #{most_recent_id}. Last fetched: #{latest_stored_id}."
-
-      if all_pages_are_fetched?
-        puts "Nothing to fetch for Timpul"
-        return
-      end
-
-      (latest_stored_id..most_recent_id).step(10) do |id|
-        fetch_single(id)
-        progressbar.increment!
-      end
-    end
-
   private
-
-    def all_pages_are_fetched?
-      latest_stored_id == most_recent_id
-    end
-
-    def latest_stored_id
-      @storage.latest_page_id
-    end
-
-    def save(page, id)
-      @storage.save(page, id)
-    end
-
-    def most_recent_id
-      @most_recent_id ||= fetch_most_recent_id
-    end
 
     def fetch_most_recent_id
       resp = JSON.parse(RestClient.get(LATEST_NEWS))
@@ -60,10 +32,6 @@ module Fetchers
 
       doc = Nokogiri::HTML(page, nil, "UTF-8")
       !doc.at_css('//h1[@itemprop="name headline"]').nil?
-    end
-
-    def progressbar
-      @progressbar ||= ProgressBar.new(most_recent_id - latest_stored_id, :bar, :counter, :rate, :eta)
     end
   end
 end
